zcash_client_sqlite: Ensure that we find out whether purely-transparent mempool transactions get mined.
This fixes an edge case where a transaction status request could expire
and be removed without having been properly serviced, for
purely-transparent transactions that were obtained by scanning the
mempool. In this circumstance, the transaction will not be associated
with a block in the course of normal blockchain scanning, because the
`scan_cached_blocks` process does not detect the transaction as being
relevant to the wallet. We should persist in querying for transaction
status so long as there is a possibility for the transaction to be
mined.