zcash_client_backend: Remove test-only `WalletRead` methods to `WalletTest`
In order to make it possible to send transparent outputs to our own
ephemeral transparent addresses to fix gap limit problems, we remove the
prohibition on sending to these addresses. This removes the last
internal use of the `WalletRead::find_account_for_ephemeral` method,
and neither this method nor `WalletRead::get_known_ephemeral_addresses`
are used by the SDKs or any known consumer of the API; since they're
powerful capabilities that are easy to misuse, these are now removed
from the public API; replacements have been added to the `WalletTest`
trait so that these APIs can continue to be used in their role in
backend tests.