zcash_client_sqlite: Correctly handle transaction expiry heights in balance & unspent note queries.
Prior to this change, transactions with unknown expiry heights were
always treated as unexpired. This could cause funds to be indefinitely
"locked" if a spending transaction was mined, detected via chain
scanning, then reorg'ed away and never mined; in this circumstance, the
wallet would see the notes and/or UTXOs spent by that transaction as
being permanently unavailable to the wallet. This change adds a
non-nullable `min_observed_height` column to the `transactions` table
that tracks the block height at which a transaction is first observed.
This change needs to be coupled to a wallet feature that continues to
attempt to retrieve such orphaned transactions for the period that the
transaction would cause a note to be locked by the orphan that spends it.