zcash_client_sqlite: Add `ON DELETE CASCADE` to `accounts.id` foreign keys.
In order to perform account deletion, we need foreign key constraints to
cascade on delete. This change adds `ON DELETE CASCADE` modifiers to
enable both account and transaction deletion.

In the process, since it is rewriting a large number of tables it cleans
up and regularizes column naming, and removes columns that are no longer
used where relevant. In the process it also simplifies a few joins in
`v_transactions` since internal transaction identifiers can be used for
joins instead of txid byte arrays.