zcash_client_sqlite: Improve resilience of transaction status checks.
This removes the use of the `request_expiry` field of the
`tx_retrieval_queue` table in favor of a more robust and consistent
strategy for checking for whether transactions have been mined. Now that
we have the observation height for a transaction, we can use comparisons
against this height to determine if we've waited long enough for
confirmation to determine that we should treat a transaction as having
expired without having been mined.