zcash_client_sqlite: Identify additional wallet-internal outputs in `store_transaction_to_be_sent`.
In the process of removing the `PaysToEphemeralAddress` error variant, I
found that some of the tests had been working around the inability to
send funds directly to an ephemeral address, and decided to simplify
this code. However, this led me to discover a number of edge cases
wherein the gap limit was not being correctly updated in response to
wallet addresses being observed in transactions. After fixing this
problem, it revealed that a number of the tests had been incorrectly
detecting changes to account balance; funds that were being sent to
internal addresses were not being correctly detected as outputs
controlled by the wallet, and were testing against the incorrect balance
as a result.