zcash_client_sqlite: Exclude wallet-internal ephemeral outputs from balance computation.
We expect all outputs to ephemeral addresses to fall into one of two
categories:
- A wallet-internal deshielding output created as part of sending to a
  TEX address.
- An output sent to us after we provided a one-time use ephemeral
  address to some third party, such as for a DEX swap or onramp
  transaction.

As such, we modify transparent balance calculation to ignore ephemeral
outputs that fall into the first category, because we do not expect
these to become part of the wallet's spendable balance; only in the case
of a partial TEX transaction failure (where the first transaction in the
pair is mined, but the second fails to broadcast and/or is never mined)
will we exclude outputs that need to be re-shielded to the wallet.

We need to make this change because we cannot yet reliably detect
purely-transparent transactions involving ephemeral addresses without
introducing privacy leaks, except in limited cases. In wallet restore
scenarios, we want to make sure that transparent functionality in the
wallet doesn't end up bricked by the wallet incorrectly treating
previously-spent UTXOs as spendable.