Remove PVZT v2 (#161)
The current version of PCZT corresponds to PCZT V2 which is not backward
compatible with PCZT V1.
In this PR, we revert the implementation back to PCZT V1 to restore
compatibility with the existing system.
The PCZT V2 implementation will be reintroduced later.

In this PR, we also update Orchard to use the latest version. This
Orchard version supports PCZT V1 only.