zcash_client_sqlite: Relax exclusions on wallet-internal ephemeral UTXOs
In zcash/librustzcash#1985, a clause was added to UTXO queries to ensure
that ephemeral outputs produced by wallet-internal transactions are
excluded from no such outputs are ever regarded as spendable. However,
this is too strict; it inhibits recovery of funds that end up "stuck" in
an ephemeral address due to the partial failure of a TEX send. This
relaxes that check such that if a TXO has actually been observed to be
unspent after the second transaction in a TEX chain would have expired,
we consider that unspentness information to be authoritative.