zcash_client_backend: Mark notes as spent in store_decrypted_tx.
Prior to this change, transactions being detected via mempool scanning
would not result in notes those transactions spend being marked as
spent; that would only happen as a consequence of calls to `put_blocks`
after the transactions in question are mined. Ordinarily this should be
indistinguishable to a mobile wallet user, but it closes a loophole that
may be relevant for Zallet.