zcash_client_backend: Modify transaction construction to use `UnifiedFullViewingKey::select_ovk`
This alters the OVK behavior for transactions created by
`zcash_client_backend::data_api::wallet::create_proposed_transaction` to
mirror the approach implemented by `z_sendmany` in `zcashd`.

Prior to this change, external outputs to a particular pool were made
decryptable by the pool-specific OVK corresponding to the UFVK for
account that provided funds to the transaction, irrespective of whether
any of the funds provided were sourced from that pool.

After this change, a single preferred OVK is selected from the UFVK for
the source account based upon the pools that provided inputs to the
transaction; that OVK is then used to encrypt all external outputs,
irrespective of pool.