zip321: Make `zip321::Payment::amount` an optional field.
Previously, the amount field of the payment would be set to zero if no
amount parameter was present; however, this behavior is not specified by
ZIP 321. The semantics of this field have now been altered such that if
no amount is specified by a payment request, this should be interpreted
to mean that the sender of the transaction should specify an amount for
the payment.